Make platform-agnostic GetSdJwtVcIssuerMetadataOps, GetSdJwtVcIssuerJwkSetKtorOps
